making emote parsing testable
